+++ trunk/GoJava7/dobritsa/KickstarterD_Web_DB/src/main/java/ua/com/goit/gojava7/kickstarter/servlet/ProjectServlet.java
@@ -48,7 +48,7 @@ public class ProjectServlet extends HttpServlet {
 			throws ServletException, IOException {
 		
 		log.info("doGet()...");		
-		int projectId = Integer.parseInt(request.getParameter("id"));
+		Long projectId = Long.parseLong(request.getParameter("id"));
 		
 		Project project = projectDao.get(projectId);
 		project.setPledged(paymentDao.calculatePledgedForProject(project.getId()));
