+++ trunk/GoJava7/dobritsa/KickstarterD_Web_DB/src/main/java/ua/com/goit/gojava7/kickstarter/servlet/ProjectsServlet.java
@@ -44,7 +44,7 @@ public class ProjectsServlet extends HttpServlet {
 	protected void doGet(HttpServletRequest request, HttpServletResponse response)
 			throws ServletException, IOException {
 		log.info("doGet()...");
-		int categoryId = Integer.parseInt(request.getParameter("id"));
+		Long categoryId = Long.parseLong(request.getParameter("id"));
 
 		List<Project> projects = projectDao.getByCategory(categoryId);
 		for (Project project : projects) {
